Workbasket list view and initial detail view.
