+++ workbaskets/jinja2/includes/workbaskets/workbasket-detail.jinja
@@ -37,7 +37,6 @@
       }) }}
 
       {% include "includes/common/pagination.jinja" %}
-
   </div>
 </div>
 {% endif %}
