+++ workbaskets/jinja2/workbaskets/list.jinja
@@ -1,22 +1,5 @@
-{% extends "layouts/layout.jinja" %}
+{% extends "layouts/list.jinja" %}
 
-{% set page_title = "Find and edit workbaskets" %}
-
-{% block beforeContent %}
-  {{ govukBreadcrumbs({
-    "items": [
-      {"text": "Home", "href": url("index")},
-      {"text": page_title}
-    ]
-  }) }}
-{% endblock %}
-
-{% block content %}
-  <h1 class="govuk-heading-xl">{{ page_title }}</h1>
-  <p class="govuk-body">
-    Enter criteria to help find a workbasket.
-    Alternatively, <a href="">create a new workbasket</a>.
-  </p>
-
-  {% include "includes/workbaskets/workbasket-list.jinja" %}
-{% endblock %}
+{% set form_url = "workbaskets:workbasket-ui-list" %}
+{% set ommit_create_link = True %}
+{% set list_include = "includes/workbaskets/workbasket-list.jinja" %}
