+++ workbaskets/urls.py
@@ -2,42 +2,43 @@ from django.urls import include
 from django.urls import path
 from rest_framework import routers
 
-from workbaskets import views
+from workbaskets.views import api as api_views
+from workbaskets.views import ui as ui_views
 
 app_name = "workbaskets"
 
 api_router = routers.DefaultRouter()
-api_router.register(r"workbaskets", views.WorkBasketViewSet)
+api_router.register(r"workbaskets", api_views.WorkBasketViewSet)
 
 ui_patterns = [
     path(
         "",
-        views.WorkBasketList.as_view(),
+        ui_views.WorkBasketList.as_view(),
         name="workbasket-ui-list",
     ),
     path(
         f"<pk>/",
-        views.WorkBasketDetail.as_view(),
+        ui_views.WorkBasketDetail.as_view(),
         name="workbasket-ui-detail",
     ),
     path(
         f"<pk>/submit/",
-        views.WorkBasketSubmit.as_view(),
+        ui_views.WorkBasketSubmit.as_view(),
         name="workbasket-ui-submit",
     ),
     path(
         f"<pk>/delete-changes/",
-        views.WorkBasketDeleteChanges.as_view(),
+        ui_views.WorkBasketDeleteChanges.as_view(),
         name="workbasket-ui-delete-changes",
     ),
     path(
         f"<pk>/delete-changes-done/",
-        views.WorkBasketDeleteChangesDone.as_view(),
+        ui_views.WorkBasketDeleteChangesDone.as_view(),
         name="workbasket-ui-delete-changes-done",
     ),
     path(
         "download",
-        views.download_envelope,
+        ui_views.download_envelope,
         name="workbasket-download",
     ),
 ]
