+++ workbaskets/views/api.py
@@ -1,0 +1,25 @@
+from rest_framework import renderers
+from rest_framework import viewsets
+
+from common.renderers import TaricXMLRenderer
+from workbaskets.models import WorkBasket
+from workbaskets.serializers import WorkBasketSerializer
+
+
+class WorkBasketViewSet(viewsets.ModelViewSet):
+    """API endpoint that allows workbaskets to be viewed and edited."""
+
+    queryset = WorkBasket.objects.prefetch_related("transactions")
+    filterset_fields = ("status",)
+    serializer_class = WorkBasketSerializer
+    renderer_classes = [
+        renderers.JSONRenderer,
+        renderers.BrowsableAPIRenderer,
+        TaricXMLRenderer,
+    ]
+    search_fields = ["title"]
+
+    def get_template_names(self, *args, **kwargs):
+        if self.detail:
+            return ["workbaskets/taric/workbasket_detail.xml"]
+        return ["workbaskets/taric/workbasket_list.xml"]
